requirements.txt
# Core dependencies
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
python-dotenv==1.0.0

# WebSocket
websockets==11.0.3
websocket-client==1.6.4

# OpenAI
openai==1.3.0

# Audio Processing
numpy==1.24.3
sounddevice==0.4.6

# Async Support
aiohttp==3.8.5

# Development
black==23.7.0
pytest==7.4.0

# CORS and Security
starlette==0.27.0
pydantic==2.4.2

# Development Tools
asyncio==3.4.3 